Move client enum logic into `codegen-client`
